Extract MnmPublicadas rows to it's own component
