+++ app/MnmPublicadas.js
@@ -2,35 +2,18 @@ import React, { Component } from 'react'
 import {
     StyleSheet,
     View,
-    PixelRatio,
-    Image,
-    Text,
     ListView,
-    TouchableHighlight,
     ActivityIndicatorIOS,
     StatusBar
 } from 'react-native'
 
+var MnmPublicadasRow = require('./MnmPublicadasRow');
 var screen = require('Dimensions').get('window');
 var moment = require('moment');
-var Icon = require('react-native-vector-icons/EvilIcons');
 var ThumborURLBuilder = require('thumbor-url-builder');
 
 import { THUMBOR_KEY, THUMBOR_URL} from './ThumborConfig'
 
-var MnmEntrada = require('./MnmEntrada');
-
-class MnmEntryDate extends Component {
-    render() {
-        var dateText = 'Publicada';
-        if (this.props.section === 'Nuevas') {
-            dateText = 'Enviada';
-        }
-        return (
-            <Text style={styles.pubDate}>{dateText} {this.props.date}</Text>
-        );
-    }
-}
 
 class MnmPublicadas extends Component {
     constructor(props) {
@@ -76,48 +59,9 @@ class MnmPublicadas extends Component {
         });
     }
 
-    rowPressed(entry) {
-        this.props.navigator.push({
-            component: MnmEntrada,
-            passProps: {entrada: entry},
-        });
-    }
-
-    renderImage(entry) {
-        if (entry.mediaPublished) {
-            return (
-                <View style={styles.imgContainer}>
-                    <Image source={{uri: entry.mediaPublished}} style={styles.image}/>
-                </View>
-            );
-        }
-    }
-
     renderRow(rowData) {
         return (
-            <TouchableHighlight
-                onPress={() => this.rowPressed(rowData)}
-                underlayColor={'#FAFAFA'}>
-                <View style={styles.rowContainer}>
-                    <View style={styles.infoContainer}>
-                        <Text style={styles.meneos}>{rowData.votes} ↑</Text>
-                        <Text style={styles.negatives}>{rowData.negatives} ↓</Text>
-                        <View style={styles.comments}>
-                            <Text style={styles.commentsText}>{rowData.comments}</Text>
-                            <Icon style={styles.commentsIcon} name='comment'
-                                size={20} color='#95a5a6'/>
-                        </View>
-                        <MnmEntryDate section={this.props.section}
-                            date={rowData.dateFromNow}/>
-                    </View>
-                    {this.renderImage(rowData)}
-                    <View style={styles.titleContent}>
-                        <Text style={styles.title}>
-                            {rowData.title}
-                        </Text>
-                    </View>
-                </View>
-            </TouchableHighlight>
+            <MnmPublicadasRow entry={rowData} navigator={this.props.navigator} />
         );
     }
 
@@ -129,13 +73,13 @@ class MnmPublicadas extends Component {
                         automaticallyAdjustContentInsets={false}
                     />;
         } else {
-          return
+          return (
             <ActivityIndicatorIOS
-          style={styles.centering}
-                        animating={true}
-                        color='#262626'
-                        size='large'
-                    />;
+              style={styles.centering}
+              animating={true}
+              color="#262626"
+              size="large" />
+          );
         }
     }
 
@@ -161,81 +105,7 @@ var styles = StyleSheet.create({
     list: {
         flex: 1,
         backgroundColor: '#FAFAFA',
-    },
-    rowContainer: {
-        flex: 1,
-        flexDirection: 'column',
-        paddingBottom: 25,
-        marginLeft: 10,
-        marginRight: 10,
-        borderBottomColor: '#BDC3C7',
-        borderBottomWidth: 1 / PixelRatio.get(),
-    },
-    infoContainer: {
-        flex: 1,
-        flexDirection: 'row',
-        alignItems: 'flex-start',
-        marginTop: 10,
-        marginBottom: 10,
-    },
-    meneos: {
-        color: '#d35400',
-        fontSize: 14,
-        fontWeight: '300',
-        marginRight: 10,
-    },
-    negatives: {
-        flex: 1,
-        color: '#e74c3c',
-        fontSize: 14,
-        fontWeight: '300',
-    },
-    comments: {
-        flex: 1,
-        flexDirection: 'row',
-    },
-    commentsText: {
-        color: '#95a5a6',
-        fontSize: 14,
-        fontWeight: '300',
-    },
-    commentsIcon: {
-        bottom: 3,
-    },
-    pubDate: {
-        flex: 5,
-        color: '#95a5a6',
-        fontSize: 14,
-        fontWeight: '300',
-        textAlign: 'right',
-    },
-    imgContainer: {
-        flex: 1,
-        height: 155,
-        width: screen.width - 20,
-        marginBottom: 10,
-    },
-    image: {
-        flex: 1,
-        position: 'absolute',
-        top: 0,
-        left: 0,
-        bottom: 0,
-        right: 0,
-        borderRadius: 3,
-        resizeMode: 'cover',
-        backgroundColor: '#222',
-    },
-    titleContent: {
-        flex: 1,
-    },
-    title: {
-        flex: 1,
-        color: '#262626',
-        fontWeight: '300',
-        fontSize: 20,
-        marginTop: 5,
-    },
+    }
 });
 
 module.exports = MnmPublicadas;
